backend/requirements.txt
Django==4.2.7
psycopg2-binary
djangorestframework
channels
channels-redis
django-cors-headers
django-oauth-toolkit==2.0.0